requirements.txt
# Core
Django==4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1

# Channels for WebSockets
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# Database
# psycopg2-binary==2.9.9  # Removed - not compatible with Python 3.13
psycopg==3.1.17  # Core psycopg package
psycopg-binary==3.2.9  # Updated to latest version available for Python 3.13

# Background tasks
celery==5.3.6
redis==5.0.1

# Environment variables
python-dotenv==1.0.1

# Testing
pytest==7.4.4
pytest-django==4.7.0

# API Documentation
drf-spectacular==0.27.0

# Utilities
# Pillow==10.1.0  # Commented out until needed
django-filter==23.5